==> Project-1-ECMO-Complications/Code/05_analysis-time-to-event-up-to-wk1.Rmd <==
---
title: "Mechanical Complications of ECMO: Analysis of Time to First Event, Up to 7 Days"
author: "Camden Lopez"
date: "`r Sys.Date()`"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      message = FALSE,
                      warning = FALSE,
                      fig.width = 7,
                      fig.height = 7)

library(coxme)
library(ehahelper)
library(knitr)
library(mice)
library(tidyverse)
source("00_helpers.R")

theme_set(theme_light())

load("ecmo.mi.RData")
```

# Pump failure

```{r}
tmp <-
  mi.bl %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(pump.fail.hr, pump.fail.event) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.bl.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()

tmp <-
  mi.hr6 %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(pump.fail.hr.after.6hr, pump.fail.event.after.6hr) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.hr6.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()

tmp <-
  mi.hr30 %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(pump.fail.hr.after.30hr, pump.fail.event.after.30hr) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.hr30.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()
```

```{r}
tbl.pump.fail <- NULL

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         age..10yr +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sbp..20mmhg +
                         dbp..10mmhg +
                         ph..0.1 +
                         pco2..20mmhg +
                         lactate..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sbp.24..20mmhg +
                         dbp.24..10mmhg +
                         ph.24..0.1 +
                         pco2.24..20mmhg +
                         lactate.24..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         cardiac.arrest +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sepsis.primary.or.with.vasoact +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.bl.wk1 %>%
                   complete(action = "long", include = TRUE) %>%
                   filter(!pre.existing.tracheostomy) %>%
                   as.mids(),
                 coxme(Surv(tstart, tstop, event) ~
                         vent..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         pf.ratio..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         pf.ratio.24..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.hr6.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         pump.flow.4 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.pump.fail <-
  bind_rows(tbl.pump.fail,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         pump.flow.24 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())
```

```{r}
tbl.pump.fail <-
  tbl.pump.fail %>%
  mutate(haz.ratio =
           sprintf("%.2f (%.2f, %.2f)",
                   estimate, ci.lower, ci.upper),
         p.val =
           sapply(p.value, format_pval))

tbl.pump.fail %>%
  select(term, haz.ratio, p.val) %>%
  kable()

tbl.pump.fail %>%
  write_csv("../Output/estimates.pump.fail.hr.upto1wk.csv")
```

\pagebreak

# Oxygenator failure

```{r}
tmp <-
  mi.bl %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(oxy.fail.hr, oxy.fail.event) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.bl.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()

tmp <-
  mi.hr6 %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(oxy.fail.hr.after.6hr, oxy.fail.event.after.6hr) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.hr6.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()

tmp <-
  mi.hr30 %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(oxy.fail.hr.after.30hr, oxy.fail.event.after.30hr) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.hr30.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()
```

```{r}
tbl.oxy.fail <- NULL

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         age..10yr +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sbp..20mmhg +
                         dbp..10mmhg +
                         ph..0.1 +
                         pco2..20mmhg +
                         lactate..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sbp.24..20mmhg +
                         dbp.24..10mmhg +
                         ph.24..0.1 +
                         pco2.24..20mmhg +
                         lactate.24..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         cardiac.arrest +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sepsis.primary.or.with.vasoact +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.bl.wk1 %>%
                   complete(action = "long", include = TRUE) %>%
                   filter(!pre.existing.tracheostomy) %>%
                   as.mids(),
                 coxme(Surv(tstart, tstop, event) ~
                         vent..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         pf.ratio..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         pf.ratio.24..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.hr6.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         pump.flow.4 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.oxy.fail <-
  bind_rows(tbl.oxy.fail,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         pump.flow.24 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())
```

```{r}
tbl.oxy.fail <-
  tbl.oxy.fail %>%
  mutate(haz.ratio =
           sprintf("%.2f (%.2f, %.2f)",
                   estimate, ci.lower, ci.upper),
         p.val =
           sapply(p.value, format_pval))

tbl.oxy.fail %>%
  select(term, haz.ratio, p.val) %>%
  kable()

tbl.oxy.fail %>%
  write_csv("../Output/estimates.oxy.fail.hr.upto1wk.csv")
```

\pagebreak

# Circuit clot

```{r}
tmp <-
  mi.bl %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(circuit.clot.hr, circuit.clot.event) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.bl.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()

tmp <-
  mi.hr6 %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(circuit.clot.hr.after.6hr, circuit.clot.event.after.6hr) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.hr6.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()

tmp <-
  mi.hr30 %>%
  complete(action = "long", include = TRUE) %>%
  survSplit(Surv(circuit.clot.hr.after.30hr, circuit.clot.event.after.30hr) ~ .,
            cut = 7 * 24,
            start = "tstart",
            end = "tstop",
            event = "event",
            episode = "episode",
            data = .)

mi.hr30.wk1 <-
  tmp %>%
  filter(episode == 1) %>%
  as.mids()
```

```{r}
tbl.circuit.clot <- NULL

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         age..10yr +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sbp..20mmhg +
                         dbp..10mmhg +
                         ph..0.1 +
                         pco2..20mmhg +
                         lactate..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sbp.24..20mmhg +
                         dbp.24..10mmhg +
                         ph.24..0.1 +
                         pco2.24..20mmhg +
                         lactate.24..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         cardiac.arrest +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sepsis.primary.or.with.vasoact +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.bl.wk1 %>%
                   complete(action = "long", include = TRUE) %>%
                   filter(!pre.existing.tracheostomy) %>%
                   as.mids(),
                 coxme(Surv(tstart, tstop, event) ~
                         vent..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.bl.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         pf.ratio..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         pf.ratio.24..log2 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.hr6.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         pump.flow.4 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())

tbl.circuit.clot <-
  bind_rows(tbl.circuit.clot,
            with(mi.hr30.wk1,
                 coxme(Surv(tstart, tstop, event) ~
                         sex +
                         weight..20kg +
                         pump.flow.24 +
                         (1 | center.id))) %>%
              coxme_pool_and_estimate())
```

```{r}
tbl.circuit.clot <-
  tbl.circuit.clot %>%
  mutate(haz.ratio =
           sprintf("%.2f (%.2f, %.2f)",
                   estimate, ci.lower, ci.upper),
         p.val =
           sapply(p.value, format_pval))

tbl.circuit.clot %>%
  select(term, haz.ratio, p.val) %>%
  kable()

tbl.circuit.clot %>%
  write_csv("../Output/estimates.circuit.clot.hr.upto1wk.csv")
```
